aggregate data aggDS aggregate data based on the tree.

aggregate data aggDS aggregate data based on the tree.

aggDS(
  TSE,
  assay = "counts",
  sample_id = "sample_id",
  group_id = "group_id",
  cluster_id = "cluster_id",
  FUN = sum,
  message = FALSE
)

Arguments

TSE
A TreeSummarizedExperiment object

assay
A name or a numeric value to index a matrix in assays.

sample_id
The column name of sample id.

group_id
The column name of group id.

cluster_id
The column name of clusters.

FUN
The function

message
TRUE or FALSE. If TRUE, the running process is shown.

Value

A singleCellExperiment object. In assays, each matrix is data of a node on the tree. Rows of matrix represent genes and columns represent samples.

Examples

library(treeclimbR)
library(ape)

tr <- rtree(3, tip.label = LETTERS[1:3])
cc <- matrix(rpois(50, 10), nrow = 5)
rownames(cc) <- paste0("gene", 1:5)
colnames(cc) <- paste0("cell", 1:10)
cd <- data.frame(sid = rep(1:2, each = 5),
                 gid = rep(letters[1:2], each = 5),
                 cid = sample(LETTERS[1:3], size = 10, replace = TRUE),
                 stringsAsFactors = FALSE)
tse <- TreeSummarizedExperiment(assays = list(cc),
                                colTree = tr, 
                                colNodeLab = cd$cid,
                                colData = cd)

out <- aggDS(TSE = tse, assay = 1, sample_id = "sid",
             group_id = "gid", cluster_id = "cid")

#> Error in showNode(tree = tree, only.leaf = FALSE, use.alias = TRUE): could not find function "showNode"