generate a count table

.doCount generates a count table given some available information. Here, the information includes the fold change between two conditions, the parameters of Negative Binomial distritubion (which the sample library size follows), a data table from real data (to estimate the proportion of each entity in the sample), and the number of samples in two different groups or conditions.

.doCount(data, FC, nSam, mu, size, n)

Arguments

data
A matrix or data frame, representing the count table from real data

FC
A numeric vector, representing the fold changes

nSam
A vector of length two containing the number of samples in two different groups or conditions.

mu, size
The parameters of the Negative Binomial distribution (see mu and size in rnbinom.)

n
A numeric value, representing the number of count tables generated. If above one, the output is a list of matrices.

Value

a matrix or a list of matrices

Author

Ruizhu Huang