Extract table with node-level DA/DS results

Extract a table with the top-ranked nodes from a DA/DS analysis output (generated by runDA or runDS).

nodeResult(
  object,
  n = 10,
  type = c("DA", "DS"),
  adjust_method = "BH",
  sort_by = "PValue",
  p_value = 1
)

Arguments

object

The output from runDA or runDS.

n

An integer indicating the maximum number of entities to return.

type

Either "DA" (for object from runDA) or "DS" (for object from runDS).

adjust_method

A character string specifying the method used to adjust p-values for multiple testing. See p.adjust for possible values.

sort_by

A character string specifying the sorting method. This will be passed to topTags. Possibilities are "PValue" for p-value, "logFC" for absolute log-fold change or "none" for no sorting.

p_value

A numeric cutoff value for adjusted p-values. This will be passed to topTags. Only entities with adjusted p-values equal or lower than specified are returned.

Value

A data frame with results for all nodes passing the imposed thresholds. The columns logFC, logCPM,

PValue, FDR, F (or LR) are from (the output table of) topTags. The node column stores the node number for each entity. Note: FDR is corrected over all features and nodes when the specified type = "DS".

Author

Ruizhu Huang, Charlotte Soneson

Examples

suppressPackageStartupMessages({
    library(TreeSummarizedExperiment)
})

lse <- readRDS(system.file("extdata", "da_sim_100_30_18de.rds",
                           package = "treeclimbR"))
tse <- aggTSE(lse, rowLevel = showNode(tree = rowTree(lse),
                                       only.leaf = FALSE))
dd <- model.matrix( ~ group, data = colData(tse))
out <- runDA(TSE = tse, feature_on_row = TRUE,
             assay = "counts", option = "glmQL",
             design = dd, contrast = NULL,
             normalize = TRUE)

## Top 10 nodes with DA
nodeResult(out, n = 10)
#>           node      logFC   logCPM         F       PValue          FDR
#> alias_102  102 -0.6706780 18.47290 230.74419 2.611542e-20 2.376503e-18
#> alias_114  114 -0.6164477 17.74557 134.01085 1.016946e-15 4.627102e-14
#> alias_115  115 -0.6363735 17.40596 122.65881 5.022099e-15 1.523370e-13
#> alias_103  103 -0.7519305 17.17217 110.95711 2.923727e-14 6.651479e-13
#> alias_116  116 -0.7205022 16.78934  80.45318 5.770607e-12 1.050251e-10
#> alias_118  118 -0.6862480 16.51284  70.25858 4.494370e-11 6.816461e-10
#> alias_110  110 -0.8660117 16.18212  62.68176 2.328574e-10 3.027146e-09
#> alias_101  101 -0.1640957 19.93063  62.05689 2.675624e-10 3.043523e-09
#> alias_112  112 -0.8746824 15.41939  44.75563 1.891109e-08 1.912121e-07
#> alias_120  120 -0.8760583 15.82848  36.17668 2.121532e-07 1.778098e-06