Table of the top differentially abundant entities

Extracts the most differentially abundant entities from a test object, ranked either by p-value or by absolute log-fold-change.

nodeResult(
  object,
  n = 10,
  type = c("DA", "DS"),
  adjust_method = "BH",
  sort_by = "PValue",
  p_value = 1
)

Arguments

object
The output from runDA or runDS.

n
A integer, maximum number of entities to return.

type
"DA" (object from runDA) or "DS" (object from runDS)

adjust_method
A character string specifying the method used to adjust p-values for multiple testing. See p.adjust for possible values.

sort_by
A character string specifying the sort method. Possibilities are "PValue" for p-value, "logFC" for absolute log-fold change or "none" for no sorting.

p_value
A numeric cutoff value for adjusted p-values. Only entities with adjusted p-values equal or lower than specified are returned.

Value

A data frame. Columns including logFC, logCPM, PValue, FDR, F (or LR) are from (the output table of) topTags. The node column stores the node number for each entities. Note: FDR is corrected over all features and nodes when the specified type = "DS".

Examples

library(TreeSummarizedExperiment)
library(treeclimbR)
set.seed(1)
count <- matrix(rnbinom(300,size=1,mu=10),nrow=10)
colnames(count) <- paste(rep(LETTERS[1:3], each = 10), rep(1:10,3), sep = "_")
rownames(count) <- tinyTree$tip.label
count[1, ] <- 0
rowInf <- DataFrame(var1 = sample(letters[1:3], 10, replace = TRUE),
                    var2 = sample(c(TRUE, FALSE), 10, replace = TRUE))
colInf <- DataFrame(gg = factor(sample(1:3, 30, replace = TRUE)),
                    group = rep(LETTERS[1:3], each = 10))
lse <- TreeSummarizedExperiment(assays = list(count),
                                rowData = rowInf,
                                colData = colInf,
                                rowTree = tinyTree)
nodes <- showNode(tree = tinyTree, only.leaf = FALSE)

#> Error in showNode(tree = tinyTree, only.leaf = FALSE): could not find function "showNode"

tse <- aggValue(x = lse, rowLevel = nodes)

#> Error in aggValue(x = lse, rowLevel = nodes): object 'nodes' not found

dd <- model.matrix( ~ group, data = colInf)
out <- runDA(TSE = tse, feature_on_row = TRUE,
             assay = 1, option = "glmQL",
             design = dd, contrast = NULL, 
             normalize = TRUE, 
             group_column = "group")

#> Error in is(TSE, "TreeSummarizedExperiment"): object 'tse' not found

             
topOut <- nodeResult(out, n = 10)

#> Error in topTags(object = object$edgeR_results, n = n, adjust.method = adjust_method,     sort.by = sort_by, p.value = p_value): object 'out' not found