Parameter estimation in Dirichlet-multinomial distribution

parEstimate is a wrapper function generated from the function dirmult with default setting on init, initscalar, epsilon, trace and mode. It allows the input obj to accept matrix or leafSummarizedExperiment and output the value pi and theta.

parEstimate(obj)

Arguments

obj
A matrix or leafSummarizedExperiment. Samples in the columns and entities in the rows.

Value

A list including “pi” and “theta”

Author

Ruizhu Huang

Examples

set.seed(1)
y <- matrix(rnbinom(100,size=1,mu=10),nrow=10)
colnames(y) <- paste("S", 1:10, sep = "")
rownames(y) <- tinyTree$tip.label

toy_tse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                    assays = list(y))
res <- parEstimate(obj = toy_tse)

#> Iteration 1: Log-likelihood value: -1758.74224075617

#> Iteration 2: Log-likelihood value: -1756.76469974966

#> Iteration 3: Log-likelihood value: -1756.57155575818

#> Iteration 4: Log-likelihood value: -1756.56834891469

#> Iteration 5: Log-likelihood value: -1756.56834779961

metadata(res)$assays.par

#> $pi
#>         t2         t7         t6         t9         t4         t8        t10 
#> 0.10448739 0.08713723 0.07954515 0.07375044 0.13113238 0.10483482 0.09405853 
#>         t1         t5         t3 
#> 0.08372658 0.08552694 0.15580054 
#> 
#> $theta
#> [1] 0.1133465
#>