Test for differential abundance using edgeR

runDA tests differential abundance of entities using functions from the edgeR. This adapts edgerWrp to accept input as a TreeSummarizedExperiment (TSE) object instead of matrix. To clarify, the count table is stored in the assays table of TSE object. Features could be stored either in rows or columns. Here, we consider features are in rows. Then, samples are in columns and the sample information is in colData. The tree that stores the hierarchical information about features is in rowTree. Each row of assays table could be mapped to a node of the tree. Data on rows that are mapped to internal nodes is generated from data on leaf nodes. Normalization for samples is automatically performed by edgeR package and the library size is calculated using features that are mapped to leaf nodes. More details are in Vignette.

runDA(
  TSE,
  feature_on_row = TRUE,
  assay = NULL,
  option = c("glm", "glmQL"),
  design = NULL,
  contrast = NULL,
  filter_min_count = 10,
  filter_min_total_count = 15,
  filter_large_n = 10,
  filter_min_prop = 0.7,
  normalize = TRUE,
  normalize_method = "TMM",
  group_column = "group",
  design_terms = "group",
  ...
)

Arguments

TSE
A TreeSummarizedExperiment object.

feature_on_row
A logical value, TRUE or FALSE. If TRUE (default), features or entities (e.g. genes, OTUs) are in rows of the assays tables, and samples are in columns; otherwise, the other way around.

assay
A numeric value to specify which table from assays is used for analysis.

option
"glm" or "glmQL". If "glm", glmFit and glmLRT are used; otherwise, glmQLFit and glmQLFTest are used. Details about the difference between two options are in the help page of glmQLFit.

design
A numeric matrix. It must be of full column rank. Defaults to use all columns of sample annotation data to create the design matrix. The sample annotation data is stored in the colData of tse when onRow = TRUE; otherwise it is in the rowData. Note: Users should check whether the default created design matrix is exactly what they want or create their own design matrix using model.matrix.

contrast
A numeric vector specifying one contrast of the linear model coefficients to be tested equal to zero. Its length must equal to the number of columns of design. If NULL, the last coefficient will be tested equal to zero.

filter_min_count
A numeric value. It's passed to min.count of filterByExpr.

filter_min_total_count
A numeric value. It's passed to min.total.count of filterByExpr.

filter_large_n
A numeric value. It's passed to large.n of filterByExpr.

filter_min_prop
A numeric value. It's passed to min.prop of filterByExpr.

normalize
A logical value, TRUE or FALSE. The default is TRUE.

normalize_method
Normalization method to be used. See calcNormFactors for more details. It could be "bonferroni", "holm", "hochberg", "hommel", "BH", or "BY". This is passed to adjust.method of topTags

group_column
The column name of group

design_terms
The names of columns from colData (if samples in columns) that are used to generate design matrix. This is ignored if design is provided.

...
More arguments to pass to glmFit (option = "glm" or glmQLFit (option = "glmQL").

Value

a list includes edgeR_results, tree, and nodes_drop.

edgeR_results

The output of glmQLFTest or glmLRT depends on the specified option.

tree

The hiearchical structure of entities that was stored in the input tse

nodes_drop

A vector storing the alias node labels of entities. These entities are filtered before analysis due to low counts.

Details

The experimental design is specified by a design matrix and provide to model via the argument design. More details about the calculation of normalization factor could be found from calcNormFactors.

Author

Ruizhu Huang

Examples

library(TreeSummarizedExperiment)
library(treeclimbR)
set.seed(1)
count <- matrix(rnbinom(300,size=1,mu=10),nrow=10)
colnames(count) <- paste(rep(LETTERS[1:3], each = 10), rep(1:10,3), sep = "_")
rownames(count) <- tinyTree$tip.label
count[1, ] <- 0
rowInf <- DataFrame(var1 = sample(letters[1:3], 10, replace = TRUE),
                    var2 = sample(c(TRUE, FALSE), 10, replace = TRUE))
colInf <- DataFrame(gg = factor(sample(1:3, 30, replace = TRUE)),
                    group = rep(LETTERS[1:3], each = 10))
lse <- TreeSummarizedExperiment(assays = list(count),
                                rowData = rowInf,
                                colData = colInf,
                                rowTree = tinyTree)
nodes <- showNode(tree = tinyTree, only.leaf = FALSE)

#> Error in showNode(tree = tinyTree, only.leaf = FALSE): could not find function "showNode"

tse <- aggValue(x = lse, rowLevel = nodes)

#> Error in aggValue(x = lse, rowLevel = nodes): object 'nodes' not found

dd <- model.matrix( ~ group, data = colInf)
out <- runDA(TSE = tse, feature_on_row = TRUE,
             assay = 1, option = "glmQL",
             design = dd, contrast = NULL, 
             normalize = TRUE, 
             group_column = "group")

#> Error in is(TSE, "TreeSummarizedExperiment"): object 'tse' not found