Test for differential state using edgeR

runDA tests differential state of entities using functions from the edgeR. This adapts edgerWrp to accept input as a SummarizedExperiment (SE) object instead of matrix. To clarify, the count table is stored in the assays table of SE object. Each matrix in assays stores data corresponding to a node of the tree. Samples are in columns and features, e.g., genes or antibodies, are in rows. The sample information is in colData. The tree that stores the hierarchical information about the assays is provided via the argument tree. assays are named with the alias labels of nodes that could be mapped to the provided tree. The experimental design is specified by a design matrix and provide to model via the argument design. Alternatively, users could specify the name of columns (from colData) that are passed to model.matrix to generate the design matrix.

runDS(
  SE,
  tree,
  option = c("glm", "glmQL"),
  design = NULL,
  contrast = NULL,
  filter_min_count = 1,
  filter_min_total_count = 15,
  filter_large_n = 10,
  filter_min_prop = 1,
  min_cells = 10,
  normalize = TRUE,
  normalize_method = "TMM",
  group_column = "group_id",
  design_terms = "group_id",
  message = TRUE,
  ...
)

Arguments

SE
A SummarizedExperiment object.

tree
A phylo object. Each matrix in assays (of SE) stores data on one node of the tree.

option
"glm" or "glmQL". If "glm", glmFit and glmLRT are used; otherwise, glmQLFit and glmQLFTest are used. Details about the difference between two options are in the help page of glmQLFit.

design
A numeric matrix. It must be of full column rank. Defaults to use all columns of sample annotation data to create the design matrix. The sample annotation data is stored in the colData of tse when onRow = TRUE; otherwise it is in the rowData. Note: Users should check whether the default created design matrix is exactly what they want or create their own design matrix using model.matrix.

contrast
A numeric vector specifying one contrast of the linear model coefficients to be tested equal to zero. Its length must equal to the number of columns of design. If NULL, the last coefficient will be tested equal to zero.

filter_min_count
A numeric value. It's passed to min.count of filterByExpr.

filter_min_total_count
A numeric value. It's passed to min.total.count of filterByExpr.

filter_large_n
A numeric value. It's passed to large.n of filterByExpr.

filter_min_prop
A numeric value. It's passed to min.prop of filterByExpr.

min_cells
A numeric value. The minimum number of cells in a node to include a node in the analysis.

normalize
A logical value, TRUE or FALSE. The default is TRUE.

normalize_method
Normalization method to be used. See calcNormFactors for more details. It could be "bonferroni", "holm", "hochberg", "hommel", "BH", or "BY". This is passed to adjust.method of topTags

group_column
The column name of group

design_terms
The names of columns from colData (if samples in columns) that are used to generate design matrix. This is ignored if design is provided.

message
A logical value, TRUE or FALSE. If TRUE, the running process is printed out.

...
More arguments to pass to glmFit (option = "glm" or glmQLFit (option = "glmQL").

Value

a list includes edgeR_results, tree, and nodes_drop.

edgeR_results

A list, each of its element is the output of glmQLFTest or glmLRT depends on the specified option. The list is named by the alias label of nodes

tree

The hiearchical structure of entities that was stored in the input tse

nodes_drop

A vector storing the alias node labels of entities. These nodes have no sufficient data to run analysis.

Author

Ruizhu Huang

Examples

library(TreeSummarizedExperiment)

data(tinyTree)

set.seed(1)
# data: genes in rows, cells in columns
ncell <- 500
ngene <- 20
dat <- matrix(rpois(ngene*ncell, lambda = 20), nrow = ngene)

# cells belong to different clusters
clus <- sample(tinyTree$tip.label, 500, replace = TRUE)
sid <- sample(1:4, 500, replace = TRUE)
gid <- ifelse(sid %in% 1:2, "A", "B")
colD <- data.frame(cluster_id = clus,
                   sample_id = sid,
                   group = gid,
                   stringsAsFactors = FALSE)
# TreeSummarizedExperiment object
d_tse <- TreeSummarizedExperiment(
            assays = list(dat),
            colData = colD,
            colTree = tinyTree, 
            colNodeLab = clus)
d_se <- aggDS(TSE = d_tse, 
              assay = 1, sample_id = "sample_id",
              group_id = "group",
              cluster_id = "cluster_id", FUN = sum)

#> Error in showNode(tree = tree, only.leaf = FALSE, use.alias = TRUE): could not find function "showNode"

res <- runDS(SE = d_se, tree = tinyTree, 
             option = "glm", group_column = "group",
             design_terms = "group")

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'assays': object 'd_se' not found

nodeResult(res, type = "DS")

#> Error in nodeResult(res, type = "DS"): object 'res' not found