Select branches

selNode selects branches meeting the specified criteria in the number of leaves and the count proportion.

selNode(
  pr = NULL,
  obj = NULL,
  data = NULL,
  tree = NULL,
  minTip = 0,
  maxTip = Inf,
  minPr = 0,
  maxPr = 1,
  skip = NULL,
  all = FALSE
)

Arguments

pr
A named numeric vector to provide proportions of entities.

obj
A TreeSummarizedExperiment object.

data
A count table with entities in rows and samples in columns.

tree
A phylo object

minTip
the minimum number of leaves in the selected branch

maxTip
The maximum number of leaves in the selected branch

minPr
The minimum count proportion of the selected branch in a sample. A value between 0 and 1.

maxPr
The maximum count proportion of the selected branch in a sample. A value between 0 and 1.

skip
A character vector of node labels. These nodes are not the descendants or the ancestors of the selected branch.

all
TRUE or FALSE. Default is FALSE. If FALSE, the branch node of a branch, which meet the requirements and has the minimum count proportion, is returned; otherwise branch nodes of all branches meeting the requirements are returned.

Value

The node whose descendant branch has the lowest proportion

Author

Ruizhu Huang

Examples

set.seed(1)
data("tinyTree")
toyTable <- matrix(rnbinom(40, size = 1, mu = 10), nrow = 10)
colnames(toyTable) <- paste(rep(LETTERS[1:2], each = 2),
rep(1:2, 2), sep = "_")
rownames(toyTable) <- tinyTree$tip.label

dat <- parEstimate(obj = toyTable)

#> Error in parEstimate(obj = toyTable): class(obj) %in% c("matrix", "list", "TreeSummarizedExperiment") are not all TRUE

(out1 <- selNode(tree = tinyTree, data = dat, all = TRUE))

#> Error in selNode(tree = tinyTree, data = dat, all = TRUE): object 'dat' not found

(out2 <- selNode(tree = tinyTree, data = dat, 
                 minTip = 4, maxTip = 9,
                 minPr = 0, maxPr = 0.8, all = TRUE))

#> Error in selNode(tree = tinyTree, data = dat, minTip = 4, maxTip = 9,     minPr = 0, maxPr = 0.8, all = TRUE): object 'dat' not found

## provide obj
lse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                assays = list(toyTable))
(out3 <- selNode(obj = lse, all = TRUE))

#> Iteration 1: Log-likelihood value: -633.210632859034

#> Iteration 2: Log-likelihood value: -631.489192200762

#> Iteration 3: Log-likelihood value: -631.155358434505

#> Iteration 4: Log-likelihood value: -631.132407245137

#> Iteration 5: Log-likelihood value: -631.132252461193

#> Iteration 6: Log-likelihood value: -631.132252452367

#>          nodeNum nodeLab proportion numTip
#> alias_11      11 Node_11  1.0000000     10
#> alias_12      12 Node_12  0.8644560      9
#> alias_13      13 Node_13  0.2148379      3
#> alias_14      14 Node_14  0.1488320      2
#> alias_15      15 Node_15  0.6496181      6
#> alias_16      16 Node_16  0.5177530      5
#> alias_17      17 Node_17  0.3484034      3
#> alias_18      18 Node_18  0.1716573      2
#> alias_19      19 Node_19  0.1693496      2

# All nodes except node 1
(out4 <- selNode(obj = lse, skip = 1, all = TRUE))

#> Iteration 1: Log-likelihood value: -633.210632859034

#> Iteration 2: Log-likelihood value: -631.489192200762

#> Iteration 3: Log-likelihood value: -631.155358434505

#> Iteration 4: Log-likelihood value: -631.132407245137

#> Iteration 5: Log-likelihood value: -631.132252461193

#> Iteration 6: Log-likelihood value: -631.132252452367

#>          nodeNum nodeLab proportion numTip
#> alias_14      14 Node_14  0.1488320      2
#> alias_15      15 Node_15  0.6496181      6
#> alias_16      16 Node_16  0.5177530      5
#> alias_17      17 Node_17  0.3484034      3
#> alias_18      18 Node_18  0.1716573      2
#> alias_19      19 Node_19  0.1693496      2